Update Step Function Definition to work with new input model
