+++ stepfunctions/secretscanner.yml
@@ -5,13 +5,13 @@ States:
   Filter Action Type:
     Type: Choice
     Choices:
-      - Variable: "$.action"
+      - Variable: "$.detail.action"
         StringEquals: "created"
         Next: Get Secret Details
-      - Variable: "$.action"
+      - Variable: $.detail.action"
         StringEquals: "resolved"
         Next: Success State
-      - Variable: "$.action"
+      - Variable: $.detail.action"
         StringEquals: "reopened"
         Next: Success State
     Default: Success State
